requirements.txt
qdrant-client
python-dotenv
google-generativeai
pypdf
scikit-learn
matplotlib
pdfplumber